Silencing warnings

+++ src/utils.cc
@@ -126,7 +126,7 @@ int32_t jump_consistent_hash(uint64_t key, int32_t num_buckets)
 		key = key * 2862933555777941757ULL + 1;
 		j = (b + 1) * (double(1LL << 31) / double((key >> 33) + 1));
 	}
-	return b;
+	return (int32_t) b;
 }
